[SymForce] Replace 1e-8 with std::numeric_limits<Scalar>::epsilon()
Currently we use a default value of `1e-8` for epsilon in `FromTangent`, `ToTangent`, `Retract`, and `LocalCoordinates` in `CLASS.h.jinja`. This is problematic, particularly for `ToTangent` and `LocalCoordianates`, as `Rot3<float>::Identity().ToTangent()` (using the default epsilon), returns `nan`s (same idea applies to `LocalCoordinates`). That is to say, `1e-8` too small of an epsilon (by about a factor of 10). It also doesn't help that for `double`, we can use an epsilon as small `1e-15` just fine (so we are using much too large of an epsilon).

This PR is meant to take a step in the right direction by instead choosing the more appropriate epsilon values returned by `std::numeric_limits<Scalar>::epsilon()` (which are `1.192e-7` and `2.220e-16` about for float and double respectively).

Create epsilon.h and Epsilon<Scalar>

Also, use Epsilon<Scalar> in a few obvious places.
A change like this was needed because, for example,
the default epsilon of 10-8 was too small for
Rot3<float>::ToTangent and could result in nans when
applied to the identity.

Regenerate files with changes to Epsilon<Scalar>

typedef kDefaultEpsilond and kDefaultEpsilonf

Explicitly instantiate kDefaultEpsilon

Shorten docstring of kDefaultEpsilon

and redirect the user to symforce/notebooks/epsilon_sandbox.ipynb

Move epsilon.h to gen from opt

This is because we do not want gen to
depend on other parts of symforce (like opt)

Regenerate Files: epsilon.h move from opt to gen