[SymForce] Replace various epsilons with kDefaultEpsilon
In particular, `std::numeric_limits<Scalar>` in `LieGroupOps<Scalar>::IsClose`, and various epsilons (mostly `1e-7`) in `geo_package_cpp_test.cc` were replaced with `kDefaultEpsilon<Scalar>`.

The idea being that if we want a unified way to handle epsilon, it will probably be easier if we start converting over to `kDefaultEpsilon<Scalar>`.