[SymForce] Update outdated comment
The comment used to say that the `FixedBundleAdjustmentProblem` was generated in `symforce/test/symforce_bundle_adjustment_example_codegen_test.py`.

That file, however, does not seem to exist. The problem is, however, generated in `symforce/test/symforce_examples_bundle_adjustment_fixed_size_codegen_test.py`, so I've updated the comment to point to that file instead.