[SymForce] Disallow creation of Keys with kInvalidLetter
I'm not actually sure why we'd want this, so if there's a good reason we
can keep it.  However, it seems to me like we'd rather throw on Key
construction than run into weird issues later.  It also means that if
you see a NULLKEY you've read it out of invalid memory :)

[Testing] Existing tests?

Topic: sym-no-invalid-keys