[SymForce] Consolidate paths into path_util (mostly)
Consolidate the job of fetching include paths, library paths, etc for
things into `symforce.path_util`.  There are two remaining places other
than that: `symforce/Makefile` and `symforce/test/Makefile`.  I think we
should do something about these, but that's going to come later.  For
now, this makes it trivial to move symforce into symforce_module, and to
see the paths we'll need to put in whatever manifest we need to generate
when we get there

[Testing] Existing tests

Topic: sym-path-util
Relative: sym-no-mc-test
Reviewers: hayk, bradley