[SymForce] Purge boost
Remove uses of boost::optional.  This means we don't have an optional
type, since std::optional is C++17 and std::experimental::optional is
not in the standard so I don't think we want to use it.  This is super not ideal, it would be
nice to either roll our own optional or copy one from somewhere

[Testing] Builds without boost

Topic: sym-no-boost
Reviewers: hayk, bradley