setup.py for sym package
Add a setup.py for gen/python so that we can `pip install gen/python`.
Not sure if this is the optimal way to do this, but it works.  I'd like
to take another pass at consolidating all of this at some point, open to
doing it before this merges but I'd plan to come back to it when I get
to streamlining the whole build/install process.

Topic: sym-python-misc-2