[SymForce] Generate manifest from CMake for Make
Generates a manifest.json with information that needs to get fed to
makefiles later, like include and link directories, and the path to
lcm-gen.  Currently I'm generating it into
`symforce/build/manifest.json`.  The issue is that it needs to be
somewhere that Python can find it - I'm not sure how to do this if CMake
isn't being run by Python except to put it under the source directory.  I
think unless someone has an obviously better idea, this is something we
should revisit when we get to making SymForce pip installable, assuming
it doesn't bite us before then.

[Testing] Existing tests

Topic: sym-manifest
Reviewers: hayk, bradley