[SymForce] Problem API
Add infrastructure for building symbolic optimization problems.  The
docstrings provide good documentation of what's going on, but a
high-level overview:  there's an OptimizationProblem class which
contains the `inputs` Values for the whole problem variables, and the
`residuals` Values for the whole problem residuals.  For big problems
(currently the only flow we support), users create multiple
SubProblems to put into an OptimizationProblem; each SubProblem has
an Inputs dataclass and a build_residuals function.  These can use
Objectives, which are logical containers for a single residual function
and Params block, but are not strictly required by the API.

This review does not generate any code or run
optimization, those are coming in later PRs.