[SymForce] Sparse factors
The C++ Optimizer machinery can now handle factors with sparse
jacobians/hessians.  I did most of this by just creating parallel types
for sparse factors, partly because it would be nice to do this with
std::variant which is C++17 (Argh!).  We could alternatively roll our
own variant or use some other form of polymorphism/dynamic dispatch to
make this nicer.  There's still one `Factor`, but it can either be
sparse or dense, and supports one version of Linearize based on which it
is (calling the wrong one will assert).

[Testing] Tested on the fixed size bundle adjustment example

Topic: sym-sparse-factors
Reviewers: hayk,bradley