[SymForce] Allow comparison of cc_sym.Key with non cc_sym.Key object
This is primarily motivated to support mypy type-stubs.

MyPy expects the `__eq__` method to be defined between the class and any `object` (and not just other instances of `cc_sym.Key`).

So, to make mypy happy, this adds an additional wrapping layer to `sym::Key::operator==` to return `False` if `other` is not a `cc_sym.Key`.

This PR also adds a corresponding test case (and an additional test case that `cc_sym.Key.__eq__` returns false if compared to another `Key` with different fields, as I noticed I was missing one from before).