[SymForce] Sparse factor cleanup
Everything not tested is broken, so add tests for sparse factors.  And
of course, I forgot to implement sparse jacobian factors, so here they
are.  You can't have a sparse jacobian factor with fixed vectors,
because then we can't deduce the size the Rhs should be, but you can
have a sparse jacobian factor with dynamic vectors.

This also splits out a bunch of the factor code into factor_utils.h

[Testing] Add unit tests

Topic: sym-sparse-factors-2
Reviewers: hayk,bradley
GitOrigin-RevId: da1ec730091942a1cc13b6a1f4da341b5d1facbe