[SymForce] Values is a MutableMapping[str, Any]
This means mypy is happy with unpacking with `**` (which actually only
requires Mapping, not MutableMapping).  For Mapping, only changes
required are adding `__iter__` and changing the signatures of `keys`,
`values`, and `items` to match the `Mapping` interface.  To make it a
MutableMapping only required deleting our implementation of `update`;
the default implementation works, and mypy thinks the signature of ours is
incompatible (not sure why).

Topic: sym-values-mapping
Reviewers: hayk,bradley,eric
Relative: sympc-desired-position
GitOrigin-RevId: 2cb1cbe4b2ab5c29b17324b164868e48fdb2ad73