[SymForce] Make symforce.cc_sym importable regardless of location of cc_sym.*.so
Previously one could not run `import cc_sym` (or something analogous) in
the independent build.

To fix this, a file called `cc_sym.py` was added to symforce which get's the location of `cc_sym.*.so`, adds it to the python path, then reexports all of its contents.

In order to get the location of `cc_sym.*.so`, the cmake variable `SYMFORCE_PY_EXTENSION_MODULE_OUTPUT_PATH` was added to the symforce manifest, whose value is then fetched with `symforce.path_util.cc_sym_install_dir`.

As a result, `cc_sym` can now be imported with `from symforce import cc_sym`, and the various places where `cc_sym` is used were modified to reflect this.

GitOrigin-RevId: b95ac83c106586aed35f91c8b49379f590f7311d