Add #hashable notation to skymarshal
This adds a `#hashable` notation to skymarshal which defines the hash function for these types in the Python implementation. It's not a good idea to use this notation widely (because the structs are mutable), but this approach is no worse than what we had before, and it's perhaps a bit cleaner.

In the future we may want to add support for immutable Python messages, which would make `#hashable` unnecessary.

GitOrigin-RevId: 9ff8cf920ba2f6d7ecb4c45424b88cab182eaff6