[SymForce] Fix matrix identity/inverse confusion
This is an attempt to fix the
`geo.Matrix.{identity|inverse|matrix_identity|matrix_inverse}`
confusion.  I think the most reasonable way to do so is for Matrix to no
longer inherit from Group, and instead just inherit from Storage.
Coming with that, it no longer overrides the methods required for those
classes.  I've then added a MatrixGroupOps and MatrixLieGroupOps
implementing matrices as a group under addition.  This still leaves the
question of what to do about `inverse` and `identity`, since we could
provide implementations for them using the matrix identity and inverse.
I've decided not to do this, because it seems like a footgun for users who
incorrectly call `cls.inverse` when they mean `ops.LieGroupOps.inverse`.
Instead, I've added an overload to `eye`, to cover both the fixed and
dynamic size cases, and have renamed the inverse to `inv`, which is what
sympy uses anyway.  `eye` certainly seems entirely reasonable to me,
since there are other python packages that use this name.  In a vacuum
`inverse` is a better function name than `inv`, but I think I prefer
`inv` here for the anti-footgun reason.

I've also considered things like making `GL` or `RNxM` subclasses of
`Matrix` which implement the group and lie group ops, but I don't think
those are really necessary and certainly not now.

Topic: sym-kill-matrix-identity
Reviewers: hayk,bradley
GitOrigin-RevId: 50208e8b86af15a3d58171c8547b5a03c45115a0