[SymForce] Run pylint
Run pylint, just on the main symforce package for now and not the rest
of symforce.

Also fix lots of good errors pylint catches.  There are still more that
we should fix IMO, but I'm leaving that as a lower-priority TODO to
tackle later, since this is already an improvement.

Adding @eric-skydio in case he has thoughts

Topic: sym-pylint
Reviewers: hayk,bradley,eric
GitOrigin-RevId: c9f4c13b852fa5fa3ea038e2d137a6e87af78c8c