[SymForce] Add all the camera cals to types_t and define TypeEnum
In order to support adding camera cals to `Values`, we need to add corresponding enums to `type_t`.

I didn't want to have a situation where we add new camera cals and have to remember to manually add them to `type_t`, so I set it to add all subclasses of `CameraCal` (note, only immediate children; subclasses of subclasses won't be added because there are no examples currently and I didn't want to go overkill).

The logic for coming up with the camera cal enum names was too complicated to want to keep in the jinja file I thought, so I pass it in from the outside.

Since we generate the lcm types in both `test/symforce_lcm_codegen_test.py` and `codegen/geo_package_codegen.py`, I put the logic in a new file, `codegen/lcm_types_codegen.py`, which can then be imported into both.

ALSO

For the same purpose, we also need `StorageOps<T>::EnumType` to be defined for `T` a camera cal type, so I also do that in this PR.

GitOrigin-RevId: 25f5827f00b3cd7546e382405e62d0e820551d67