[SymForce] Register CameraCal with LieGroupOps in python
Since we want to add the camera cal classes to `Values`, and `Values` requires that stored elements implement `LieGroupOps`, we want to register `CameraCal` with LieGroupOps. Note, `CameraCal`'s can be thought of as Lie groups if their storage space can be thought of as a vector space (the group operation is the vector sum of the storage, and the identity element would then be the `CameraCal` who's storage is the zero vector).

Now, we could just implement `CameraCalStorageOps`, `CameraCalGroupOps`, and `CameraCalStorageOps` (as we do with `Matrix` and `Scalar`), but this would duplicate the logic in `MatrixXXXOps` and `ScalarXXXOps`, as both `Matrix` and `Scalar` are vector spaces in the same sense as `CameraCal`. Thus, doing so would only unnecessarily bloat our codebase.

Instead, we can use the fact that when written in terms of their storage ops, all three of these classes have identical group and Lie group ops implementations.

Specifically, I created `AbstractStorageOps`, an abstract base class where all the methods of a `StorageOps` impl are abstract methods, `VectorSpaceGroupOps`, a class which implements the group ops in terms of the abstract storage methods, and `VectorSpaceLieGroupOps`, which inherits from `VectorSpaceGroupOps` and also implements its methods in terms of the abstract storage ops.

This way, to implement `LieGroupOps` for `Scalar`, `CameraCal`, and `Matrix`, you only need to implement the parts where they differ, namely, provide implementations for the abstract storage methods of `AbstractStorageOps`.

For `ScalarLieGroupOps`, this is just directly done. Since both `Matrix` and `CameraCal` inherit from `Storage`, we can fully provide implementations of `LieGroupOps` by inheriting from `ClassStorageOps` in addition to `VectorSpaceLieGroupOps`. I did this and called the class `VectorClassLieGroupOps` (which replaces `MatrixLieGroupOps`). I then registered `CameraCal` with `VectorClassLieGroupOps`.

One thing to point out, we no longer have `ScalarGroupOps`, `ScalarStorageOps`, or `MatrixGroupOps` classes. We didn't use them (except in inheritance) so I haven't replaced them.

GitOrigin-RevId: edd357675f3978c49d74d79f00d47cfa7984291d