Added better support for indexing dataclasses stored in Values objects
This fixes an error I was getting when sending Values to/from storage that contained dataclasses which stored sequences. This was an issue because we previously weren't storing the storage dimension of the sequences contained in the dataclass. Thus, when reconstructing the Values object from its index, we would try to reconstruct the dataclass element using `from_storage` + the dataclass type alone, rather than an instance of the dataclass (this is necessary when the dataclass has a sequence as a field, as stated in the `DataclassStorageOps` docstring).

GitOrigin-RevId: 6c754dd914e326d41b25d6f5dd601ccd66942ead