Added support for pickling and unpickling Values objects
Added support for pickling and unpickling Values objects. The main weird thing is the call to `.to_numerical()` that is made when pickling the object; here I'm assuming we only want to pickle numerical Values, not Values containing symbolic objects (this would be a bit more tricky, and I'm not sure there's a convincing use-case for this just yet).
﻿

GitOrigin-RevId: 49ea0924fc13033315c6ca41a98b8855ede58af7