Add option to prune the values object used for optimization.
Internally, the optimizer will copy this object multiple times. When
this object is large, this can consume a substantial amount of compute.

Perhaps the ideal change would be to avoid making these copies in the
optimizer, but that would be a trickier/larger change to make.

Topic: prune_values
Relative: best_linearization
GitOrigin-RevId: e3b416bf701dec7decdcef0011ea80f215a66b50