Lazily print code in Codegen
We only need to print code when generate_function
is called.

However, we don't always want to call generate_function
on every Codegen object. For example, if a Codegen
object is generated so that we can call with_linearization
on it.

Since print code is relatively expensive, it's better
to only call it if we actually intend to use the results.

Topic: lazy_print_code_evaluation
GitOrigin-RevId: b708676d062debfe6ce6b6ff2c7779f4c6b29962