[SymForce] Move python ops templates to ops dir
Was in geo_package.

This is to make it easier to reuse the templates
in cam specifically, and more generally to make
it easier to use templates from a variety of
courses (partially decoupling the location of
the jinja template from the location of the
generated file).

Topic: rearrange_template_locations
Relative: openable_in_template_util
GitOrigin-RevId: 3ac8bbc23fd07e94d13770db90c6e13effcf734d