[Symforce] Added ops support for NoneTypes so that Dataclasses can have optional fields
This is used in the case where e.g. we want to call `to_storage` on a Dataclass with an optional field (i.e. one that defaults to `None`). This is necessary because calling `to_storage` on a Dataclass results in calling `to_storage` on each element of the dataclass, which currently raises an error if one of the fields is set to `None`.

An alternate implementation would be to handle `NoneType` in the dataclass StorageOps, but I figure this is a more general solution. I think this is less confusing to read since we would otherwise need to check whether each element of the dataclass is `None` before performing ops on it. Using the setup in this PR, we can just use ops on each element of a dataclass without having to add `if element is not None` everywhere.

GitOrigin-RevId: a727840478ebf7640657e05eef2448de0457039d