[Symforce] Better variable naming in opt factor construction
While playing around with constructing a large optimization problem, I
was finding it difficult to debug errors resulting from mis-constructing
the inputs to the optimization problem. Steps that caused the error:

1) I mistakenly constructed an element of the input Values as a scalar
rather than a geo.V4 object
2) When calling the linearization function with the incorrectly
constructed inputs, and exception is raised when we try to access the
i-th element of the geo.V4 object because was is passed as a scalar by
mistake.

To see the line raising the error, I commented out
`python_util.remove_if_exists(codegen_data["output_dir"])` in
`factor.py`. However, this wasn't helpful with the current
implementation because each input to the linearization function is named
"arg{i}", of which there can be hundreds.

This change results in each argument having a more readable name.

Topic: symforce_opt_problem_arg_rename
GitOrigin-RevId: 70f9fac8be9753781f138d2fab32f1725237c0dd