[SymForce] Speed up assert in Codegen construction
This was previously taking like 30% of the time for
symforce_gen_codegen_test.  The bottleneck is entirely the call to
free_symbols.  This is much faster if we call it once, by assembling
everything into a symengine matrix first.

symforce_gen_codegen test goes from 40s to 26s

Reviewers: hayk,bradley
Topic: sym-assert-fast
GitOrigin-RevId: 6c2ecc38f027c0d25db13642515e5cfaa821b70b