[SymForce] Replace sympy.count_ops w/ better impl
In this commit, I just copy the existing sympy
implementation into symforce.

The only changes made to it are:
- Adding typing to the signature
- Reformatting
- Adding # type: ignore to various places
- Adding # pylint: disable=... to various places

I opted to do it this way to be minimally invasive to
make it easier to identify which changes we've made vs
what was already present.

Topic: improve_sympy_count_ops
GitOrigin-RevId: a37529f77803e0991aadb31d4afba94029547396