[SymForce] Add property setters for DoubleSphereCC
Currently you have to index into the distortion_coeffs
vector. But this is not as readable because it's not
clear in which order alpha and xi are stored.

Plus we already have property getters for these values,
so it's kind of odd that we don't have setters.

Topic: double_sphere_setters
GitOrigin-RevId: 29e7ff6485353740c48dcc1476a8e8dbe7b3ec2e