[SymForce] Rework camera cal constructors
In order to make the constructors consistent and clear
between C++, Python, and the synbolic versions (for all
the camera cal classes), we are changing the indivdually
named distortion coefficients to have their own arguments.

For example, omega in ATANCameraCal will be a named argument
of the __init__ method, and the same goes for alpha and xi
of DoubleSphereCameraCal.

In order to ensure every camera cal class can still be created
from just a focal_length, principal_point, and distortion_coeffs,
the factory classmethod from_distortion_coeffs has been added.

The arguments to this class must be numerical if the class being
called in a PolynomialCameraCal and SphericalCameraCal, as they
can not calculate their critical_undistorted_radius and
critical_theta, respectively, from symbolic values.

ATANCameraCal was given __init__ and symbolic overloads because
it has the named distortion coefficient omega.

DoubleSphereCameraCal was given __init__ and symbolic overloads
because it has the named distortion coefficients xi and alpha.

PolynomialCameraCal was given a from_distortion_coeffs overload
because it has fields beyond focal_length, principal_point, and
distortion_coeffs. It already had __init__ and symbolic overloads.

The same applies for SphericalCameraCal.

Topic: camera_cal_init_rework
Relative: double_sphere_setters
GitOrigin-RevId: 81f04b750d9cff0320d962dace7f54659ee5e006