[Symforce] Added support for loading factors from files
Gives the ability to reuse a previously generated linearization
function. Allows for easier analysis of factors that have already been
generated, as we don't have to regenerate the factor every time we e.g.
restart a jupyter notebook.

Also changes `optimized_keys` in `factor.py` to be a list instead of a
set to avoid accidental reordering of the keys.

Topic: symforce_factor_reload
GitOrigin-RevId: 7b5a49c145aa1f69e825140db0b5e394881694bb