[SymForce] Diffs in test failures
Print the diffs for test failures, should make it easier to diagnose
things in CI for instance.  Using [difflib](https://docs.python.org/3/library/difflib.html)
which is built into python3.  Not just passing the `msg` argument to
`assertMultiLineStringEqual` because there doesn't seem to be a way to
compute that lazily on failure.

Reviewers: hayk,bradley
Topic: sym-failure-diffs
GitOrigin-RevId: fbf62b5f1216dce79847fdb60ad0f34df8fa5548