[SymForce] Allow geo.M.__setitem__ with np.float64
Specifically, the problem arrises if the key is
a tuple and you are assigning a 1d list of np.float64.

The symenginepy code relied on a TypeError being raised
if you attempt to index into matrix value (because that
is what you would ordinarily expect), but np.float64
for some reason instead raises an IndexError (even though
that's usually used for an out of bounds access). The
result is that the symenginepy code breaks, because an
unexpected exception is raised.

This can be fixed by simply adding an IndexError to the
list of exceptions it expects.

When the backend is sympy, things break for unrelated
reasons. Need to address that as well (and probably
will as I imagine the issue is blocking something I'm
interested in doing).

Topic: symenginepy_setitem_fix
GitOrigin-RevId: ec64278bc886ebf376940e02a6edcb3e1f160339