[symforce] add dont_flatten_args to _get_subs_dict
Pretty much all the time spent in codegen for optimization problems is
in _get_subs_dicts for the subs in `codegen_util:format_symbols`

In that particular place, we already have a flat list of scalar symbols
to sub.

This adds a kwarg to subs that will just use the passed in subs dict
without any of the type checking or flattening we otherwise do.

In testing this sped up codegen for opt problems by a factor of ~2x

topic: symforce-unsafe-subs
relative: symforce-factor-noformat
GitOrigin-RevId: 0d212ed5bcce900048f31e43aafc71808c99df29