[SymForce] Consistent Matrix storage order
It is column major in C++/Eigen and row
major in our symbolic python representation.

To make it consistent, I'm changing the python
storage order to be column major.

Nothing needs to be done for python sym, because
there we do not define a to_storage for matrices.

Topic: matrix_storage_order
GitOrigin-RevId: 4027467efbad3b566728dac6e9d58f1421f3a70b