[symforce] add an option to symengine wrapper to skip eval on _sympy_
the latest bottleneck in codegen seems to be when we convert back to
sympy for code printing when using the symengine backend.

this change adds an option to skip eval when construction sympy objects
in symengine wrapper eval methods. testing on examples shows pretty
significant speedups (90s -> 20s) as a representative test case.

Not on by default since it does increase ops by 1 or 2 for most test
cases: seems like we're missing out on simplifications for max/min.

topic: symforce-skip-eval
GitOrigin-RevId: dcacc4d9f0863011adb434f1a20a8def0bd682f8