[SymForce] Replace Matrix arg with CSCFormat arg
in jinja macro sparse_matrix_init.

This was because element wasn't really doing much,
and by replacing it with a CSCFormat object, we're
able to offload much of the duty spec was playing in
the function onto the CSCFormat, reducing how much
the reader needs to know about Codegen objects in
order to understand the macro.

Topic: replace_matrix_with_cscformat
Relative: remove_uneeded_none_check
GitOrigin-RevId: 44a3142835b88e6b513d7350012f4bdcdf3659e8