[SymForce] Make sparse_matrix_init more readable
The fact that there was so much duplication and the lack
of spacing between the branches made it confusing to
sort out what exactly the macro is doing. Hopefully this
is easier for other people to understand.

Also, broke up a long line.

Topic: readable_sparse_matrix_init
Relative: replace_matrix_with_cscformat
GitOrigin-RevId: 588313cd90e2c8fc412fd07d5730d0224edf2ea4