[Symforce] Allow explicit template instantiation for C++ codegen
This change adds an option to generate functions which are declared in a
header file and defined in a `.cc` file. This is typically useful when
dealing with large generated functions that result in long compile
times, or in cases where the user may want to compile the generated
function with custom compiler flags that are only applied to the
generated function.

Topic: symforce_explicit_template_instantiation
Relative: symforce_cpp_opt_problem_gen
GitOrigin-RevId: bfd35ed7789085b6100835c7a7dfbdefeb71fc57