[SymForce] Remove unused codegen_util code
`codegen_util.perform_cse` is never called with
`substitute_inputs = True` (somewhat surprising,
considering that's the default arg).

The `input_symbols` argument is only needed if
`substitute_inputs` is `True`, and so there's no
reason to keep it around either.

This should hopefully simplify `codegen_util.py`
and make it easier for future maintainers to
maintain.

Also, the docstring of `perform_cse` was pretty
outdated (the claim that some of the outputs were
`XXX.values_recursive()` was just wrong) and
unhelpful, so I've updated it.

Topic: remove_unused_perform_cse_code
GitOrigin-RevId: 416bb3e646a5698a5e30648fe73ae81d77633ee1