[SymForce] Factor out py residual wrapper - cc_sym
This PR addresses a few issues.

1)

Principally this reduces the repition of logic between
cc_sym.Factor and sym::Factord.

The problem was that both have two sets of constructors:
one for when one list of keys is passed in, and one for
when two lists of keys is passed in.

What is done when only one list is passed in is to call
the version for two lists, using the one list for both.

This logic was duplicated in cc_sym. This could pose a
problem if we changed what should be done for the one
key case for sym::Factor, but forgot to update
cc_sym.Factor.

This PR addresses that by calling the overloads of
sym::Factor directly for each case.

To support this, the wrapping of the python function
to have the required C++ signature was pulled out into
a seperate function to be called multiple times.

2)

A related issue is that we only test the one key case
in symforce_cc_sym_test.py. So I've improved the test
coverage a bit by making sure we can at least call
the list constructor from python.

3)

The last issue is less of a big deal, but because the
one list and two list versions for cc_sym had the
same names (MakeHessianFactor), we needed some extra
boilerplate to disambiguate which function pointer we
meant by &MakeHessianFactor.

One option would have been to move the implementation of
MakeHessianFactor to a lambda in place, but I was afraid
this would overcrowd the class definition for Factor.

So instead, I just gave the two versions different names
`MakeHessianFactorCommonKeys` for one list,
`MakeHessianFactorSeparateKeys` for two lists.
This way no help is needed to disambiguate which function
pointer we want.

Topic: factor_one_list_two_list
GitOrigin-RevId: c9504509e134dd5ba52d5d9c79403959a7cbb4f3