[SymForce] Fix return type: str -> sm.Symbol
_get_scalar_keys_recursive used to say its
return type is T.List[str] when it actually
returns a T.List[sm.Symbol] (a shame that
we currently don't catch that sort of thing
yet with type checking).

Anyway, this can be very confusing to a reader
of code using this function, as it will lead
them to believe the returned value is the wrong
type in the calling code (speaking from experience).

Also, fixing a typo in a comment while I'm at it.
treate -> treat

Topic: fix_return_type
GitOrigin-RevId: a7718e93e6ee280fc0dbaf76f323550c0c501d75