[SymForce] Add sym::Factor::SparseJacobianFunc
Before this commit, there was no good way to construct
construct a `sym::Factor` from a function mapping a
`sym::Values` to a residual and sparse jacobian (but not
hessian or rhs).

One could manually wrap the function into a
`SparseHessianFactor` and use the normal `Factor` constructor,
but:
a) I didn't think of that before making this commit.
b) That would require you (the user) to know how to construct
   a hessian from the jacobian (either by knowing internal
   functions to call), or do it yourself (requiring you to
   know that it's only necessary to compute the lower triangle
   of the hessian and the most efficient way to compute JtJ).

To make this easier, this commit adds the `SparseJacobianFunc`
type (and renames `JacobianFunc` to `DenseJacobianFunc`), makes
the code to construct a `Factor` from a `DenseJacobianFunc` a
bit more generic, then reuses that code to construct a `Factor`
from a `SparseJacobianFunc`.

Also, since the `DenseJacobianFunc` constructor wasn't tested
before this, this commit also adds a test of that constructor
(in addition to the test of the constructor from a
`SparseJacobianFunc`).

Topic: sparse_jacobian_func
Relative: 3_fold_hessian
GitOrigin-RevId: 3d30564d9a999264a2d4c9ca018b7043ce55455b