[SymForce] Give generate_function return type
Previously, the return type was just a map from strings
to T.Any. This was inconvenient, as it makes it rather
difficult to know what the values actually being
returned are (it's not in the doc string, so you have
to read through the source code or run it in the interpreter).

To support having one common type for regardless of target language,
I've change the `"python_function_dir"` and `"cpp_function_dir"` to
both be `"function_dir"`. Perhaps something more elaborate could
be done to support the different names, but I figured it would
be better to keep it simple.

Most of the commit is changing the places where it is used.

Also, changed all the paths to be of type `pathlib.Path` (there were
some `str`s in there before).

Topic: generate_function_return_type
GitOrigin-RevId: b4eb98162f02cb95dd22c6f81964687a40e82445