[Symforce] Allow specific explicit template instantiations
Previously we allowed users to set a flag to indicate if they wanted
generated C++ functions to be explicitly instantiated, and would
explicitly instantiate a float and a double version of the function.
However, compliling very large functions can be slow because both the
float and double versions of the function must be compiled, even if only
one is used. This PR allows the user to do explicit template
instantiation on whatever types they like, meaning they can choose to
only do explicit template instantiation on float or double if they like.

Topic: symforce_explicit_template_instantiation_list
GitOrigin-RevId: 27e5b54441a80928b85019df44fc605ecbdf75aa