[SymForce] Enable sparse cc_sym.Factor creation
Before this commit, it was not possible to create a wrapped
sparse `sym::Factor`.

This commit enables the creation of wrapped sparse `sym::Factor`s
by adding the sparse argument to
 - `cc_sym.Factor`,
 - `cc_sym.Factor.Jacobian`

If sparse is True, then a sparse factor is created. If sparse
is False, then a dense factor is created. The sparse argument
is defaulted to False (i.e., to create a dense factor).

If sparse is True and the python residual function does not
return the jacobian/hessian as a `scipy.sparse.csc_matrix`, a
value error is raised when `cc_sym.Factor.linearize` is called.

Similarly, is sparse if False and the python residual does
return the jacobian/hessian as a `scipy.sparse.csc_matrix`, a
value error is also raised when `cc_sym.Factor.linearize` is called.

This is not strictly necessary, as sparse matrices can be converted
to dense matrices and vice versa, but this will add an overhead,
and the user would likely rather have an exception raised then to
be silently (and unwittingly) converting between dense and sparse
on every optimization iteration.

It might seem preferable to automatically deduce whether a sparse
or dense factor should be created based on the return type of the
python residual function, but this is not possible (without going
great lengths perhaps) because there is no way to deduce the
return type of a python function without calling it.

An alternative might have been to have a special python residual
wrapper class whose type signals whether or not to create a sparse
or dense factor, but that seems rather cumbersome for the user and
not different in essence from just having a sparse argument.

Another alternative would have been to have a different named
constructor for the dense and sparse versions, but that seemed
like it would be too much clutter in the namespace.

Last note, I replaced the two versions of WrapPyHessianFunc (and
WrapPyJacobianFunc) for one key vs two keys with a variadic template.
This was unrelated to the core change, but I think makes the code
more readable. The change was made for no other reason. It also
doesn't require function pointer disambiguation, which is nice.

Topic: sparse_cc_sym_factor
Relative: sparse_jacobian_func
GitOrigin-RevId: 7f7abf1a2236d5a9ff27639ce37ab146bad1601d